Add an import statement

+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/visualize/DOTCFGVisualizer.java
@@ -33,6 +33,7 @@ import org.checkerframework.dataflow.expression.FieldAccess;
 import org.checkerframework.dataflow.expression.LocalVariable;
 import org.checkerframework.dataflow.expression.MethodCall;
 import org.checkerframework.javacutil.BugInCF;
+import org.checkerframework.javacutil.TreeUtils;
 import org.checkerframework.javacutil.UserError;
 
 /** Generate a graph description in the DOT language of a control graph. */
@@ -221,7 +222,7 @@ public class DOTCFGVisualizer<
             CFGLambda cfgLambda = (CFGLambda) ast;
             String clsName = cfgLambda.getSimpleClassName();
             String methodName = cfgLambda.getMethodName();
-            int uid = TreeUtils.treeUids.get(cfgLambda.getCode());
+            long uid = TreeUtils.treeUids.get(cfgLambda.getCode());
             outFile.append(clsName);
             outFile.append("-");
             outFile.append(methodName);
